+++ database/migrations/2016_08_21_900008_create_lead_table.php
@@ -13,16 +13,16 @@ class CreateLeadTable extends Migration
     public function up()
     {
         Schema::create('leads', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('user_id')->unsigned();
-            $table->bigInteger('client_id')->unsigned()->nullable()->default(null);
-            $table->bigInteger('status_id')->unsigned();
-            $table->bigInteger('reason_id')->unsigned()->nullable()->default(null);
+            $table->increments('id');
+            $table->integer('user_id')->unsigned();
+            $table->integer('client_id')->unsigned()->nullable()->default(null);
+            $table->integer('status_id')->unsigned();
+            $table->integer('reason_id')->unsigned()->nullable()->default(null);
             $table->string('client', 255)->nullable()->default(null);
             $table->string('contact', 255)->nullable()->default(null);
             $table->string('source', 255)->nullable()->default(null);
             $table->string('needs', 255)->nullable()->default(null);
-            $table->bigInteger('budget')->nullable()->default(null);
+            $table->integer('budget')->nullable()->default(null);
             $table->text('background')->nullable()->default(null);
             $table->text('gdrive')->nullable()->default(null);
             $table->enum('dealstatus', ['unknown', 'win', 'lose'])->nullable()->default('unknown');
