+++ database/migrations/2016_08_21_900024_create_worksheet_table.php
@@ -13,9 +13,9 @@ class CreateWorksheetTable extends Migration
     public function up()
     {
         Schema::create('worksheets', function (Blueprint $table) {
-            $table->bigIncrements('id');
-            $table->bigInteger('project_id')->unsigned();
-            $table->bigInteger('worker_id')->unsigned();
+            $table->increments('id');
+            $table->integer('project_id')->unsigned();
+            $table->integer('worker_id')->unsigned();
             $table->text('task');
             $table->text('description');
             $table->dateTime('start');
@@ -25,7 +25,6 @@ class CreateWorksheetTable extends Migration
             $table->decimal('amount', 10, 2);
             $table->char('currency', 3);
             $table->boolean('billable');
-            // $table->char('hash', 32)->unique();
             $table->timestamps();
         });
     }
